requirements.txt
flask==3.1.2
flask_pymongo==3.0.1
python_dotenv==1.1.1
pymongo==4.14.1
scikit-learn==1.7.1
pandas==2.3.2
bcrypt==4.3.0